requirements.txt
Flask==2.3.3
flask-cors==4.0.0
tensorflow==2.15.0
Pillow==10.3.0
numpy==1.24.4
